aiopyvuebot/templates/notebot/requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Telegram bot (async, modern)
aiogram>=3.7.0

# MongoDB async driver
motor>=3.3.0

# Environment variables
python-dotenv

# YAML for string localization
pyyaml
